Only add violating properties to the message
