+++ test/test_jsonschema_draft1.rb
@@ -19,6 +19,8 @@ class JSONSchemaDraft1Test < Minitest::Test
 
   include ObjectValidation::AdditionalPropertiesTests
 
+  include StrictValidation
+
   include StringValidation::ValueTests
   include StringValidation::FormatTests
   include StringValidation::DateAndTimeFormatTests
